Memory improvements
Performance; internal pointers replaced with offsets (converted to pointers just-in-time), reducing memory overhead, minimum alloc sizes, possibly improving cache efficiency

Safety changes:
- Custom size and offset types added, matching the footprint chosen for the allocator

- operator+ on array handles modified to return a specialized subset type (with bounds checking & an explicit reference to the superset array) instead of the weird child arrays generated up to now

The new & old solutions both have the issue where a superset/parent array can be destroyed without the child's awareness, but I'm hoping that will be relatively tractable under the array subset approach; unlike before, where child arrays were basically fully separate handles that happened to share the parent's handle ID and have a nonzero [dataOffset]

- Memory views (byte spans) introduced, as a neater alternative to address-of operator hackery (like &*handle), or type punning with ArrayAllocHandle<uint8_t> and handle ID copies (how GetBytesHandle() worked).

Actually getting bytes out of spans (ByteSpan::Bytes(...)) requires writing out the span's length, and that has improved memory safety in the rest of the codebase by itself, since it subconsciously prompts to check that span sizes match up with buffers/mapped ranges before performing copies.

- Much greater assertion coverage, probably bordering on excessive, but hopefully catches some issues

Removed MemMgrVerification test suite (massive potential timesink, distracts from actual rendering; I think I'll focus on assertions and defensive allocator choices going forward)

+++ SandboxApp/SandboxApp.cpp
@@ -113,7 +113,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance,
         // during this call
         // MAY be worthwhile having a specialist upload class - depends how those uploads are instrumented, lots of
         // complicated thoughts there
-        frameConstants->timeSeconds = std::chrono::steady_clock::now().time_since_epoch().count() * 1e-9;
+        frameConstants->timeSeconds = static_cast<float>(std::chrono::steady_clock::now().time_since_epoch().count() * 1e-9);
         rndr->UpdateFrameConstants(frameConstants);
 
         // Draw current frame
